Few fixes for worldedit changes
