+++ worldedit-core/src/main/java/com/sk89q/worldedit/world/registry/BundledItemData.java
@@ -75,7 +75,7 @@ public class BundledItemData {
         GsonBuilder gsonBuilder = new GsonBuilder();
         gsonBuilder.registerTypeAdapter(Vector3.class, new VectorAdapter());
         Gson gson = gsonBuilder.create();
-        URL url = ResourceLoader.getResource(BundledItemData.class,"items.json");F
+        URL url = ResourceLoader.getResource(BundledItemData.class,"items.json");
         if (url == null) {
             throw new IOException("Could not find items.json");
         }
